Add unfinished position encoder module
This module still needs some work. I'm a bit confused as to how position encodings are done in https://arxiv.org/pdf/1706.03762.pdf.

Particularly, I'm not sure what the difference between $i$ and $pos$ is.

I'm also trying to figure out if it's possible to deduce $d_\text{model}$ from the input, or if it needs to be passed to `__init__`.